Sender now holds onto size in a private member variable, to avoid having it go out of scope before the send completes.
This resolves the failure of the single-PE "auto-communication test"
in tstSend_Receive on Cielito.  This test has typically passed on
other machines, probably due to lucky buffering by MPI.  MPI_Isend
will not block waiting for a matching receive, but it's not required
to buffer data.  The send buffer can only be safely reused (or
destroyed, by any mechanism) after the send completes; i.e., after a
Wait or Test.