Adding a capability to choose to add an additional set of starting directions which are zero-weight, outwardly directed angles on a level that are opposite the inwardly directed angles usually used for the starting direction calculation in curvilinear coordinates.
The "enforce exact symmetry" flag for the sweeper is now a bool and is
set to be false by default. This change requires that many of the application
test input files be modified to turn this flat on because the base files were
calculated with the previous default of being turned on.