Adding a very basic utility to interactively provide info in IPCRESS files.
This commit adds the executable "GandolfInterpreter" for installation in
draco's /bin directory so we can find out the materials, temperatures,
densities, and basic absorption opacity data that lives in an IPCRESS file.
I'm sure it can be improved upon, but it will hopefully be useful to other
people, too. The usage is
> {DRACO_INSTALL}/bin/GandolfInterpreter file.ipcress
upon which some basic prompts let you print some of the info in the file.
It's currently only for gray and multigroup absorption opacities.