+ Changing all references from uint64_t to unsigned long long in c4.
The reason for this is due to a standards clash best seen in MPI_Traits.hh.
Although we'd like to insist on uint64_t (ANSI C 99), we are only guaranteed
unsigned long long in the MPI standard. So fundamentally we have to pass using
the trait from the following function in the MPI traits:

template<>
struct MPI_Traits<unsigned long long>
{
    static MPI_Datatype element_type() { return MPI_UNSIGNED_LONG_LONG; }
};

We can only rely on short <= int <= long <= long long -- hopefully
int64_t fits into one of those boxes in a parallel call.