If Random123 was found, build and run unit and known-answer tests.
The known-answer tests emit "PASSED" on success and "FAILED" on
failure.  The unit tests almost all use assert to report failures,
which CTest seems to detect and report as "OTHER_FAULT"; the one
exception is ut_ars, which writes "error" and some informative text to
stderr.

Several of the tests, and some of the Random123 headers themselves,
generated warnings under GCC with our preferred flags (typically about
unused variables or unused-but-set parameters).  For warnings from the
Random123 headers, I followed KT's suggestion and used the GNU
system_header pragma to suppress them, in any header file that
directly included a warning-producing Random123 header.