Set -fp-model precise for all build types
tstisFinite (in ds++) expects 0.*infinity to be NaN, and therefore not
finite; this is only true with -fp-model precise or strict.  Set this
for all builds (i.e., in CMAKE_C_FLAGS), instead of relying on the
default behavior in debug builds.  (Presence of -strict-ansi seems to
set -fp-model precise, while absence of -strict-ansi seems to fall
back to the default -fp-model (fast=1)).