Adjust Intel compiler flags.
Check availability of -xHost, and add it if it's available, to enable
instructions supported by the local machine.

For Release and RelWithDebInfo flags:

Switch -fp-speculation strict to -fp-speculation safe, to match the
default when -O0 is specified.

Remove -ftz; it's the default already.

Remove -inline-level=1 (enable inlining when inline keyword is
specified, or according to C++ rules), to revert to the default,
-inline-level=2 (enable inlining at compiler's discretion).