Add -debug inline-debug-info to Intel RelWithDebInfo builds.
This option "generates enhanced debug information for inlined code",
according to the icpc man page; it associates symbols from inlined
functions with the source of the called function (i.e., the inlined
function), rather than with the caller.

Intel 13 enables this option by default when -g and an optimization
flag are both specified; Intel 12 seems to leave it off by default.
The HPCToolkit manual recommends enabling it, and it does improve
attribution of calls to inline functions.