+ Cleanup, spelling mistakes, and a commented-out proposed opacity model.
Our current analytic polynomial opacity model does not contain the
[1-exp(-\nu/T)]  correction for stimulated emission.  We could easily modify
the current model to include the correction or not (as is done with the "stem"
option in TOPS) or simply add another model.  This would provide an in-house
capability to model continuum bound-free and free-free absorptions without
lines (which reduces to Kramers' opacity law with Rosseland or Planck
weighting).

I also propose that we simplify the design in the cdi_analytic directory with
respect to analytic multigroup/ODF opacity models.  Once we provide an analytic
model, a set of multigroup and/or ODF bounds, and an averaging procedure, we
ought to be able to instantiate a first-class Analytic_MultigroupOpacity and/or
Analytic_Odfmg_Opacity object directly instead of deferring these to the
"nGray" models that (confusingly) call "getOpacity" on potentially frequency-dependent
analytic models without frequency arguments. The "Psuedo_Line" classes could
continue to inherit from these without any difficulty, too. Example coding for how
to do this already exists in Pseudo_Line_Analytic_MultigroupOpacity::getOpacity,
in Jayenne as Parser::build_MultigroupOpacity, and in Capsaicin (although
the Polynomial implementation may be incorrect there).