+ Adding "getTestInputPath" to UnitTest for better IDE integration. + Going back to the standard (non-C++11) generator until we can get Mavericks.
For a while I've been tolerating test failures with tests that read an input
file from an assumed location that is different from the subdirectory that
Xcode puts it in.  With regular Makefiles, the test input and test binaries
reside in the same directory; on Xcode, the test binaries live in
subdirectories like "Debug", "Release", etc.; the failures occur when the
tests can't find their input files one level up.  This fixes that.

I've also struggled for too long with getting Xcode 5 + C++11 working and
will put that on the back burner until we can actually upgrade to the latest
OS.