Modifications to parser package meant to more cleanly separate parser code from other code and to better support inheritance and reentrancy in class parsers.
First, I have add a function template to parser/utilities.hh:

    template<class MyClass> SP<MyClass> parse_class(Token_Stream &);

This will normally be called with code of the form

        SP<MyClass> my_object = parse_class<MyClass>(tokens);

This is meant to eliminate the use of static members of classes for parsing, such as Quadrature::parse, so that parsing code is not entangled with the classes being parsed. It is also consistent with all the other parse functions in utilities, such as parse_real or parse_integer.

No default implementation is provided for this template function. It is meant to be specialized for every class for which we desire to provide a parser. However, there is a suggested implementation, described below.

This is a relatively straightforward change to propagate. In Draco, only the Quadrature class is affected. The change is more extensive in Capsaicin, but I have propagated the change and will check that in at the same time I check in the Draco changes.

Second, I have eliminated Class_Parser.hh and Class_Parser.i.hh and replaced them with Class_Parse_Table.hh. The old Class_Parser template defined a class template that was basically a collection of static members and tables, and this was not amenable either to reentrancy or inheritance. The new Class_Parse_Table.hh  does not define a class parser class template at all, but provides a templatized parser function:

    template<class Class_Parse_Table>
    SP<typename Class_Parse_Table::Return_Class>
    parse_class_from_table(Token_Stream &tokens);

The class Class_Parse_Table must provide certain member and data functions, which are described in detail in Class_Parse_Table.hh, but the developer is otherwise given considerable flexibility in writing this class. By convention, the Class_Parse_Table class providing parsing support for MyClass is named Class_Parse_Table<MyClass>. This class template is declared but not defined in Class_Parse_Table.hh, the idea being that developers will specialize it for each class for which they want a parser.

To help developers do this, I have added template__parser.hh and template__parser.cc to the draco/environment/templates directory. If you are still using emacs for editing, you will find that the Draco->New file ... menu now includes "New C++ class parser" as an option, which generates MyClass__parser.hh and MyClass__parser.cc defining Class_Parse_Table<MyClass> for whatever MyClass you choose.

So how is parse_class_from_table intended to be used? As follows:

 template<>
 SP<Class> parse_class(Token_Stream &tokens)
 {
   return parse_class_from_table<Class_Parse_Table<Class> >(tokens);
 }

This actually appears in a comment in Class_Parse_Table.hh, from which you can conveniently cut it and paste it into your parser code, filling in Class with your class name. So why not make this the default definition? It turns out that if such a default definition is visible, then every time you call parse_class<AnyClass>, the compiler wants to see a definition of Class_Parse_Table<AnyClass>, which is highly undesirable. We want the compiler to know that parse_class<AnyClass> exists, but not try to instantiate it anywhere but in AnyClass__parser.cc, where you decide if the above definition is really what you want.