Reinstate changes to Assert, with two changes that will make it more palatable:
1. The DBC macros substitute check_cookies for the conditional test plus toss_cookies only if DBC has the 16 bit set. 
2. The check_cookies function takes character pointer rather than string arguments, which should make the call a lot less costly.

The intent is that DBC should include the 16 bit when doing coverage analysis, to eliminate the numerous untaken (and hopefully untakable) branches in the DBC macros. This will make the branch coverage statistic a bit more useful for us.


Added a test of the check_cookies function to tstAssert.cc