Update the metrics_report script to use new regression paths.
The metrics_report script is run once per month to generate an email that goes
to EAP leadership and the draco team.  It presents the current Lines-of-code and
Coverage metrics. I recently updated regression system to append the text
'-develop' to the path used for draco nightly regressions (this is the feature
branch name).  Teach the metric script to look for metric data in this location
(ccscs7:///scratch/regress/cdash/draco/Nightly_gcc-develop/...).  Also update
the path filter used by BullseyeCoverage to use the new paths.

I also slightly changed the cosmetic formatting of the report.
This commit fixes #14