Implementation of polyhedron meshes through RTT mesh format.
The "polyhedron" element types assume an 3D AMR mesh based on a hexahedron
with zero or more of its 6 faces being a "refined" face. A refined face is
one with 4 additional nodes on the centers of the 4 edges of the face and
one more additional node in the center of the face. This is the (already
extant) QUAD_9 element. There is an example mesh that has been added that
exercises this new element type, along with the QUAD_9 and QUAD_5 elements
that the polyhedron element comprises. More work will be done to support
additional types of faces that may appear in an AMR mesh cell, but this
initial implementation and test is the first step in the sense that the
exmaple mesh is a single cube with one refined face.