Adding QUAD_6 and QUAD_6o mesh elements and a correspoding RTT format test with a QUAD_6 element.
The QUAD_6 (quad6 or quad6a in RTT format) has two midpoint nodes on
adjacent sides of a quad. The QUAD_6o element as the two midpoint nodes
located on opposite sides.

A test was added that exercises the QUAD_6 element through an RTT
polyhedron mesh cell having such elements.