Provide a warning suppression for clang-3.9.
This change is needed by Jayenne to suppress a clang-3.9+ warning. The offending
code is valid C++ that is considered difficult to understand. In particular, a
static data member of a template class is initialized in a .cc file instead of
the .hh file. Because we use explicit instantiation with partial specialization,
this initialization must occur in the .cc file to allow the code to compile on
all supported platforms. Thus, I have disabled the warning.