Add optional context argument to parse_class template function.
A second signature of the parse_class template function has been added that takes a second argument of type Class_Parse_Table::Context_Type const &. This allows a context object of class-dependent type to be passed to the parser which may provide context to change its behavior.

In a similar fashion, a second version of the parse_class_from_table template function has been added that takes a second Class_Parse_Table::Context_Type & argument and passes it as an argument to the Class_Parse_Table constructor.

A test has been added to exercise this option and ensure that it works.