Add option to UnitTest to suppress messages for successful tests in UnitTest::check.
By default this option is off, so existing behavior is unchanged. When this option is turned on, messages are generated by UnitTest::check only if the bool argument is false (meaning the test failed.) This can be extremely useful for debugging a test executable that has many check calls in it.