Ensure timely deallocation of parser resources.
Ensure that destructor of Class_Parse_Table<Quadrature> releases the parsed quadrature in a timely manner. Previously the parsed quadrature would not be released until general end of execution cleanups or until another quadrature was parsed.

Provide a way to free the global pointer to the internal unit system. I'm not sure why this is implemented as a raw pointer rather than a shared_ptr, but the ability to free this storage manually is useful even if it were implemented as a shared_ptr.