Extend c4::Processor_Group::sum to operate on any random access container.
The random access container must conform to the C++11 conventions for such a containter. In particular, the value_type must be defined and &x[0] must return a pointer to T pointing to the beginning of the raw storage underlying the container.

This is needed to support use of Teuchos::ArrayView and Teuchos::ArrayRCP in the Capsaicin project.