Add capability to extend debug options set.
New functions added to allow applications to add highly application-specific options to the debug mask.

The immediate need is for Capsaicin to have a debug option to look for cycles in an input mesh. This is specific enough to Capsaicin that I do not want to introduce it as a general debug option. However, I have been thinking about making the debug option set extensible for some time in anticipation that a situation like this would come up.

We may wish to consider dropping some "standard" options if there is consensus that they are too application-specific.