Modify regression settings to support slurm on knl.
Along with the transition of MOAB to Slurm on trinitite, the consultants have
also changed the way jobs are bound to KNL processors and have turned on
hyperthreading on the KNL nodes.  This was causing our regressions to crash
because 272 MPI ranks worth of tests were starting simultaneously.  Too many
jobs running simultaneously and the fact that srun is not distributing the work
correctly is causing many tests to time out.

This commit does not fix the underlying problem but it does provide a temporary
work around.

1. Avoid a possible hang during the configure step related to 'discovering' MPI.
2. When on the KNL, only run 8 MPI ranks worth of tests simultaneously.

Refs #903 (Redmine)