Fix plot2d tests that were failing on snow.
Why are these changes needed:

Snow has an older version of xmgrace than what is found on moonlight and the
output format of the metadata section of the plot file wasn't matching the gold
files. I reworked the logic in the tstPlot2D test to allow this test to pass on
both moonlight and snow by omitting the whole metadata section of the plot file
from the comparison check. Since the size of the meta data block varies between
machines, I added logic to discover how many lines needed to be removed.

What was changed:

+ Updated the `FindGrace.cmake` script to deal with LANL HPC's directory
  structures (especially on snow):
  - The installation of xmgrace on snow uses symlinks to point to actual file
    locations. CMake was replacing symlinks with real paths too soon so that if
    draco was configured on the front end, the files couldn't be found when
    compiling on a back-end node.
  - Totally rework the FindGrace.cmake module to be more conformal to other
    'mature' cmake modules.  This rewrite also fixes the problem of resolving
    real filepaths too soon.
+ Update bench files by removing all of the metadata at the top of the file.
+ Update the `checkOutputFiles` function in `tstPlot2D.cc`
  - After the test generates a plot file, figure out how many lines are meta
    data and strip those lines from the file. The amount of data removed was
    previously hard coded (i.e.: the top 58 lines).