Add an access function for the azimuthal order for the Chebyshev-Legendre product quadrature. Add DBC checks using this function.
Tweak construction of ordinates for an Ordinate_Space so that starting directions are not requested for Cartesian geometry. Starting directions are used in curvilinear geometries to provide "initial conditions" for angle derivative terms in the gradient operator and are not needed in Cartesian geometry. The request for starting conditions for Cartesian geometry is mostly benign, but it turns out to be problematic for the Gauss-Seidel capability in 1-D geometry in the Capsaicin library, where knowing whether there are any starting directions is helpful for matching up pairs of ordinates.