Fix LLVM build warnings.
When using LLVM clang to compile Draco on ccs-net machines, the CPP macro
`#ifdef __GNUC__` was true (in addition to `#ifdef __clang__`).  This caused a
logic error for the CPP macros that push, suppress warning and pop for GCC and
clang.  This problem was fixed by changing the `#ifdef __GNUC__` checks to
`#if defined(__GNUC__) && !defined(__clang__)`.

+ References #306
+ Fixes #309