Tweak diagnostics in expression parser.
When a variable is not recognized, include the variable name in the message.
Change the rather obscure "unit incompatibility for -" to "unit incompatibility for - operator" since '-' has so many meanings.

These came up when I was trying to diagnose a bad input file, and I found the messages less useful than I would have liked for determining what was going on.