Extend Debug_Options
Add a DEBUG_PROBLEM option, which turns on code that calculates characteristics of the problem specification itself *other* than whether the problem is well-posed. For example (and this is the application I need this for) the code might estimate the minimum optical depth across a processor domain or the overall scattering fraction of a problem.

Add a DEBUG_NONE option to indicate that no debugging options are desired in a way that conveys more to a human reader than 0U does.