Improve error diagnostics for Parse_Table.
Add code such that a Parse_Table will list the known keywords for the client whenever it encounters an unknown keyword. The keyword may optionally be followed by a description. This is more likely to be useful than the users' manual that no one ever reads.

In support of this capability, add a description member to the Keyword struct. Add a default constructor for Keyword and a constructor compatible with the usual brace initializers we have been using. The latter adds the description with the description defaulting to nullptr. This allows all our existing braced initializations of Keywords to continue working as before without generationg a warning, thanks to new C++11 syntax that allows braced initializers to be treated as constructor calls.

Also in support of this capability, add a comment member to the various Token_Stream children. This is similar to the report function, but does not precede everything with a token location, which is useful for extended comments from the parser.

Exercise all these functions in the test suite.