Add a check_all function to the UnitTest class tree.
This is used for a parallel test; one that must pass on all processors and is checked synchronously.
Only one pass/fail message will be generated rather than one per processor. The failing processors will be listed if the test fails.
This makes for cleaner recovery if a test fails on only some processors, particularly for the fatal failure case.

Note: I can't think of a clean way to test the fatal failure case automatically. (I've checked it manually.) This means there will be a line of uncovered code introduced by this merge.