Add a parallel output stream.
Define a class opstream and a global object of this class, pout, corresopnding to cout. This class behaves like
any other output stream in the C++ stream library, with full support for formatting of output, manipulators,
unformatted output, &c. However, output inserted by any MPI rank into the stream is buffered and does not
appear on the console (cout) until the send() function is called on pout. At this point, all the buffered output
for rank 0 appears at the console, followed by all the output for rank 1, and so on. There is absolutely no
requirement that the output from any rank be similar to that from any other rank, or that a rank produce any
output at all. However, the send() function *must* be called synchronously.

This class is meant to have some of the functionality of the old Spin_Lock class, which was designed to serialize
sections of code. The most common use case turned out to be serializing output, which Spin_Lock *did not actually
achieve*. For this reason, Spin_Lock was abandoned.

The new opstream class is not useful except for serializing output, but it actually succeeds in this, and this
is a very important use case for code serialization.