Remove global pout.
Cannot declare a global object of this type, due to requirement that it be created after MPI is initialized and destroyed before MPI is shut down.  The user must create the stream himself at the appropriate place in his code and ensure it goes out of scope before MPI finalizes. I considered adding code to require that only one such stream can be created at a time, but realized that multiple such objects is not actually a problem due to the explicit synchronization of send().