A few more tweaks to the release scripts.
Background:

We are having a lot of trouble the the Cray environment during this release
cycle.  Unloading modules is causing errors and aborting builds.  I think I have
worked through all of the issues and updated the 'release_cray.sh' script.

Also - the new python scripts in Jayenne require python 2.7+.  The default
python on Toss2 (pinto, wolf) is 2.6 and so my initial release on these machines
was defective.  I've updated the build/run environments for toss machines to
load python/3.6.

Changes:

+ Cray machines:
  + Require the build environment on Crays to be 'clean' when the release script
    is started.  This is done by aborting if the environment variable
    ParMETIS_ROOT_DIR is defined.
  + Ensure that the 'craype-mic-knl' module is not loaded when attempting to
    unload other Cray PrgEnv modules.
  + Ensure that our TPL modules are unloaded before attempting to unload Cray
  PrgEnv modules.
+ Toss2 machines:
  + Ensure that the python/3.6 module is loaded to properly support Jayenne.
+ If a test fails, run it a second time.  This seems ridiculous, but there are
  so many system issues right now that we need this information.